Add a trait for backends to implement
This is just a copy of the public API we have right now. Later we might refine
it. We wrap the inner type dynamically (instead of e.g. using generics) so we
don't change the kind of struct that is shown publicly.

This also takes us back to a situation where the public API is documented in one
place rather than in the backend.