Introduce HidApi::disable_device_discovery().
The issue with HidApi::new_without_enumerate() is that it makes both it
and HidApi::new() potentially panic, with no way out in library
contexts.

Instead, on platforms that need it, the same effect can be achieved by
calling HidApi::disable_device_discovery() before any HidApi contexts
are created in application code.

The intention is to eventually remove HidApi::new_without_enumerate().